requirements.txt
langchain-astradb
langchain
langchain-openai
datasets
pypdf
python-dotenv
flask
pandas
torch
transformers

-e .